plugin: Defer RPC calls until the JSON-RPC socket is ready
The plugin grpc interface has to start up before the JSON-RPC socket
is available, otherwise the signer could not intervene to complete the
startup. However that also means that early grpc calls that translate
to JSON-RPC calls would end up not finding the socket. This wrapper
wraps the cln-grpc service and waits for the socket file to appear
before forwarding the call.